Snapshot testing for R objects

A flexible wrapper around testthat::expect_snapshot_file() that allows snapshotting any R object, not just data.frames. This function provides a convenient interface for snapshot testing with sensible defaults for serialization.

Usage

expect_snapshot_object(x, name, writer = save_rds, ...)

Arguments

x

An R object to snapshot. Can be any R object including lists, models, data.frames, vectors, etc.

name

character snapshot name (file extension will be added automatically)

writer

function Function to write the object to a file. Default is save_rds(). Other options include save_json(), save_deparse(), save_csv(). Custom writer functions should accept x and return a file path.

...

Arguments passed on to testthat::expect_snapshot_file

cran

Should these expectations be verified on CRAN? By default, they are not, because snapshot tests tend to be fragile because they often rely on minor details of dependencies.

compare

A function used to compare the old and new results. Defaults to compare_file_binary(), which compares files line-by-line. To compare binary files byte-by-byte, use compare_file_binary().

variant

If not-NULL, results will be saved in _snaps/{variant}/{test}/{name}. This allows you to create variants for experiments, or if the results are different on different operating systems. Snapshot variant is set by the local_reproducible_output() function and its friends. If you're developing a package you can set it in the tests/testthat/helper.R file.

Value

[NULL] (from testthat::expect_snapshot_file())

Examples

if (FALSE) { # \dontrun{
# Snapshot a list (using RDS format by default)
expect_snapshot_object(list(a = 1, b = 2, c = "text"), name = "config")

# Snapshot a model
model <- lm(mpg ~ wt, data = mtcars)
expect_snapshot_object(model, name = "model")

# Snapshot with JSON format (for human-readable diffs)
expect_snapshot_object(iris[1:5, ], name = "iris_sample", writer = save_json)

# Snapshot with deparse format
expect_snapshot_object(list(x = 1:5), name = "simple_list", writer = save_deparse)
} # }